# public/api-docs-v1.yaml
openapi: 3.0.3
info:
  title: Laravel REST API
  version: 1.0.0
servers:
  - url: http://localhost:60080/api/v1
security:
  - bearerAuth: []
paths:
...
  /users/me/password:
    post:
      tags:
        - CurrentUser
      summary: Update current user's password
      operationId: updateCurrentUserPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - new_password
              properties:
                password:
                  type: string
                  minLength: 6
                  description: Must match the current password (`current_password` rule).
                new_password:
                  type: string
                  minLength: 6
                  description: Must be different from `password` (`different:password` rule).
      responses:
        '201':
          description: Updated (empty body)
          content:
            application/json:
              schema:
                type: object
                example: {}
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
...
components:
  schemas:
...
    Error401:
      type: object
      properties:
        message:
          type: string
      example:
        message: Unauthenticated.
    Error422:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      example:
        message: The given data was invalid.
        errors:
          email:
            - The email has already been taken.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT